+++ Repositories/Interfaces/IRecordatorioRepository.cs
@@ -2,15 +2,15 @@
 
 public interface IRecordatorioRepository : IRepository<Recordatorio>
 {
-    Task<RecordatorioInsertDTO> CreateRecordatorioAsync(int idUsuario, RecordatorioInsertDTO insertDTO);
+    Task<RecordatorioInsertDTO> CreateRecordatorioAsync(long idUsuario, RecordatorioInsertDTO insertDTO);
 
-    Task<RecordatorioWithGroupDTO> CreateRecordatorioInGroupAsync(int idUsuario, GrupoDTO grupo, RecordatorioInsertDTO insertDTO);
+    Task<RecordatorioWithGroupDTO> CreateRecordatorioInGroupAsync(long idUsuario, GrupoDTO grupo, RecordatorioInsertDTO insertDTO);
 
     Task<RecordatorioUpdateDTO> UpdateRecordatorioAsync(RecordatorioUpdateDTO updateDTO);
 
     Task<RecordatorioDeleteDTO> DeleteRecordatorioAsync(long idRecordatorio);
 
-    Task<IEnumerable<RecordatorioDTO>> GetAllRecordatoriosAsync(int idUsuario);
+    Task<IEnumerable<RecordatorioDTO>> GetAllRecordatoriosAsync(long idUsuario);
 
-    Task<RecordatorioDTO> GetRecordatorioAsync(int idUsuario, Expression<Func<Recordatorio, bool>> predicate);
+    Task<RecordatorioDTO> GetRecordatorioAsync(long idUsuario, Expression<Func<Recordatorio, bool>> predicate);
 }
